% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmnnCorrect.R
\name{SMNNcorrect}
\alias{SMNNcorrect}
\title{SMNN}
\description{
This function SMNNcorrect is designed to perform supervised batch effect correction for scRNA-seq data by first identifying nearest neighbors (NNs) within corresponding clusters (or cell types) and then leveraging information from these NNs.
It takes as input raw expression matrices from two or more batches and a list of the unified cluster labels (output from unifiedClusterLabelling).
It outputs batch-corrected expression matrix for each batch.
}
\usage{
SMNNcorrect(..., batch1.cluster.labels, batch2.cluster.labels, num.defined.clusters=1, correct.others=FALSE, k=20, sigma=1, cos.norm.in=TRUE, cos.norm.out=TRUE, var.adj=TRUE, subset.row=NULL, order=NULL, n.jobs=NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{...}{Two or more expression matrices each corresponding to one batch, where each row corresponds to a gene, and each colname correspond to a cell. 
The number and order of rows should be identical across all maxtices (i.e., all batches should have the exact same gene set and in the same order).}

  \item{batch1.cluster.labels}{is a vector specifying the cluster labels of each cell from batch1. Cells not belonging to any clusters should be set to 0.}

  \item{batch2.cluster.labels}{is a vector specifying the cluster labels of each cell from batch2. Similarly, cells not belonging to any clusters should be set to 0.}

  \item{num.defined.clusters}{specifies the number of clusters matched between two or more batches. Default is 1.}

  \item{correct.others}{is a Boolean variable that defines whether to search nearest neighbors among the cells not belonging to any clusters. Default is FALSE, that is, cells not belonging to any clusters will not be considered as candidate nearest neighbors.}

  \item{k}{defines the maximum number of nearest neighbors to be identified. Default is 20.}

  \item{sigma}{defines the bandwidth of the Gaussian smoothing kernel used to compute the correction vector for each cell. Default is 1.}

  \item{cos.norm.in}{is a boolean variable that defines whether to do cosine normalization on input data before computing distances between cells.
Default is "TRUE".}

  \item{cos.norm.out}{is a boolean variable that defines whether to do cosine normalization on output data before computing corrected expression results.
Default is "TRUE".}

  \item{var.adj}{is a Boolean variable that indicates whether to do variance adjustment on the correction vectors. Default is "TRUE".}

  \item{subset.row}{NULL,}

  \item{order}{is an vector defining the reference batch and the order of the other batches to be corrected.}

  \item{n.jobs}{specifies the number of parallel jobs. It would be set to the number of cores when \code{n.jobs = NULL}.}
}
\value{
SMNNcorrect returns the following:

\itemize{
  \item corrected expression matrix for each batch
  \item information regarding NNs between the current batch under correction and the reference batch
  }
}
\references{
Yuchen Yang, Gang Li, Huijun Qian, Yun Li. SMNNcorrect 2018
}
\author{
Yuchen Yang <yyuchen@email.unc.edu>, Gang Li <franklee@live.unc.edu>, Huijun Qian <hjqian@live.unc.edu>, Yun Li <yunli@med.unc.edu>
}
\examples{
# Load the example data data_SMNN
data("data_SMNN")

# Provide the marker genes for cluster matching
markers <- c("Col1a1", "Pdgfra", "Ptprc", "Pecam1")

# Specify the cluster labels for each marker gene
cluster.info <- c(1, 1, 2, 3)

# Call function unifiedClusterLabelling to identify the corresponding clusters between two batches
matched_clusters <- unifiedClusterLabelling(data_SMNN$batch1.mat, data_SMNN$batch2.mat, features.use = markers, cluster.labels = cluster.info, min.perc = 0.3)

# Set python version used in SMNNcorrect
library(reticulate)
use_python("/nas/longleaf/apps/python/3.5.1/bin/python3")

# Perform batch effect correction using SMNNcorrect
corrected.results <- SMNNcorrect(data_SMNN$batch1.mat, data_SMNN$batch2.mat, batch1.cluster.labels = matched_clusters[[1]], batch2.cluster.labels = matched_clusters[[2]], num.defined.clusters = 3, k=20, sigma=1, cos.norm.in=TRUE, cos.norm.out=TRUE)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
